Extract video component from promo
