+++ src/components/videoPromo/index.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable react/no-unknown-property */
 // Global
 import { classnames, TTailwindString } from '@/src/common/types/tailwindcss-classnames';
-import LiteYouTubeEmbed from 'react-lite-youtube-embed';
 import ButtonLink from '../common/ButtonLink';
+import YouTube from '../common/YouTube';
 // Components
 
 type VideoProps = {
@@ -26,39 +26,6 @@ const Video = ({ videoUrl }: VideoProps): JSX.Element => (
     </video>
   </div>
 );
-
-const YouTube = ({ youTubeId }: VideoProps): JSX.Element => {
-  return (
-    <>
-      <style jsx>{`
-        /* Post-click styles */
-        .yt-lite.lyt-activated {
-          cursor: unset;
-        }
-        .yt-lite.lyt-activated::before,
-        .yt-lite.lyt-activated > .lty-playbtn {
-          opacity: 0;
-          pointer-events: none;
-        }
-      `}</style>
-      <div className={classnames('flex', 'flex-col', 'justify-center', 'w-full')}>
-        <LiteYouTubeEmbed
-          id={`${youTubeId}`} // Default none, id of the video or playlist
-          adNetwork={false} // Default true, to preconnect or not to doubleclick addresses called by YouTube iframe (the adnetwork from Google)
-          params={`loop=1&playlist=${youTubeId}`} // any params you want to pass to the URL, assume we already had '&' and pass your parameters string
-          aspectHeight={9}
-          aspectWidth={16}
-          poster="maxresdefault" // Defines the image size to call on first render as poster image. Possible values are "default","mqdefault",  "hqdefault", "sddefault" and "maxresdefault". Default value for this prop is "hqdefault". Please be aware that "sddefault" and "maxresdefault", high resolution images are not always avaialble for every video. See: https://stackoverflow.com/questions/2068344/how-do-i-get-a-youtube-video-thumbnail-from-the-youtube-api
-          title="YouTube Embed" // a11y, always provide a title for iFrames: https://dequeuniversity.com/tips/provide-iframe-titles Help the web be accessible ;)
-          noCookie={true} //Default false, connect to YouTube via the Privacy-Enhanced Mode using https://www.youtube-nocookie.com
-          wrapperClass="yt-lite bg-no-repeat bg-cover relative"
-          iframeClass="top-0 w-full h-full"
-          playerClass="lty-playbtn z-10"
-        />
-      </div>
-    </>
-  );
-};
 
 const VideoPromo = ({
   title,
